:recycle: Move custom queryset into own py module
